requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
streamlit-autorefresh>=0.3.1
plotly>=5.15.0
Pillow>=10.0.0 